requirements.txt
azure-core
azure-identity
azure-mgmt-compute
azure-mgmt-core
msal
typing-extensions